+++ demo/lib/dark_theme_page.dart
@@ -20,9 +20,9 @@ class DarkThemePageState extends ExamplePageState {
       tabs.add(
           TabData(text: 'Tab $i', content: Center(child: Text('Content $i'))));
     }
-
-    TabbedWiew tabbedView = TabbedWiew(
-        controller: TabbedWiewController(tabs), theme: TabbedViewTheme.dark());
+    TabbedWiewController controller = TabbedWiewController(tabs);
+    TabbedWiew tabbedView =
+        TabbedWiew(controller: controller, theme: TabbedViewTheme.dark());
     return tabbedView;
   }
 }