+++ demo/lib/mobile_theme_page.dart
@@ -15,9 +15,9 @@ class MobileThemePageState extends ExamplePageState {
       tabs.add(
           TabData(text: 'Tab $i', content: Center(child: Text('Content $i'))));
     }
-    TabbedWiew tabbedView = TabbedWiew(
-        controller: TabbedWiewController(tabs),
-        theme: TabbedViewTheme.mobile());
+    TabbedWiewController controller = TabbedWiewController(tabs);
+    TabbedWiew tabbedView =
+        TabbedWiew(controller: controller, theme: TabbedViewTheme.mobile());
     return tabbedView;
   }
 }